+++ CHANGES.md
@@ -1,5 +1,6 @@
 ## Optimizations:
 
+- When client closes connection, retract any existing matchmaking request
 - Instead of dumping the whole candidate list, add it to the graph as soon as it is recieved.
 - Implement a dedicated sessions manager
 - Handle deadlocks and create locks for event routing and handlers